pyproject.toml
[tool.poetry]
name = "voice-assistant"
version = "0.1.0"
description = ""
authors = ["Mohammed Shokr <mohammed.younis.elsawi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
groq = "^0.18.0"
python-dotenv = "^1.0.1"
pipecat-ai = "^0.0.56"
boto3 = "^1.36.16"
daily-python = "^0.14.2"
openai = "^1.61.1"
websockets = "^14.2"
cartesia = "^1.3.1"
tiktoken = "^0.8.0"
onnxruntime = "^1.20.1"
pypdf = "^5.3.0"
pyyaml = "^6.0.2"
modal = "^0.73.42"
python-multipart = "^0.0.20"
pypdf2 = "^3.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
